feat: add edge config implementation for cmdk tooltip
