+++ src/app/layout.tsx
@@ -3,7 +3,6 @@ import "@/common/styles/globals.scss";
 import type { Metadata, Viewport } from "next";
 import Script from "next/script";
 import { Analytics } from "@vercel/analytics/react";
-import { getAll } from "@vercel/edge-config";
 import dynamic from "next/dynamic";
 
 import { TRPCReactProvider } from "@/common/tools/trpc/react";
@@ -14,8 +13,7 @@ import TooltipProvider from "@/common/providers/TooltipProvider";
 import { inter, poppins } from "@/common/fonts";
 import { env } from "@/env";
 import { CSPostHogProvider } from "@/common/providers/analytics/providers";
-import EdgeConfigProvider from "@/common/providers/EdgeConfigProvider";
-import { edgeConfigSchema } from "@/server/ecfg/config";
+import { EdgeConfigProvider } from "@/common/providers/EdgeConfig";
 
 const PostHogPageView = dynamic(
   () => import("@/common/providers/analytics/PostHogPageView"),
@@ -77,27 +75,6 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  // * FUTURE NOTE *
-  // if edge requests count nears the threshold,
-  // we should consider caching the edge config & revalidate every 24h
-  const edgeConfigRaw = await getAll();
-  const validateEcfg = edgeConfigSchema.safeParse(edgeConfigRaw);
-  let edgeConfig;
-  if (validateEcfg.success) {
-    edgeConfig = validateEcfg.data;
-  } else {
-    console.warn(
-      "Failed to parse edge config:\n",
-      validateEcfg.error.message,
-      "\n\n",
-      "Received config:\n",
-      edgeConfigRaw,
-    );
-    // used strictly as a fallback if edge config
-    // somehow returns an unexpected format
-    edgeConfig = (await import("@/server/ecfg/config.json")).default;
-  }
-
   return (
     <html lang="en" className={`${inter.variable} ${poppins.variable}`}>
       <CSPostHogProvider>
@@ -107,7 +84,7 @@ export default async function RootLayout({
             <AuthProvider>
               <TRPCReactProvider>
                 <TooltipProvider>
-                  <EdgeConfigProvider edgeConfig={edgeConfig}>
+                  <EdgeConfigProvider>
                     <CoreLayout>{children}</CoreLayout>
                   </EdgeConfigProvider>
                 </TooltipProvider>
