+++ src/common/providers/EdgeConfig/index.ts
@@ -1,0 +1,1 @@
+export * from "./EdgeConfigProvider";
