+++ src/server/ecfg/config.json
@@ -1,3 +1,4 @@
 {
-  "test": true
+  "enableAnnouncementBanner": true,
+  "enableCmdkTooltip": true
 }
