+++ src/server/ecfg/config.ts
@@ -1,7 +1,8 @@
 import { z } from "zod";
 
 export const edgeConfigSchema = z.object({
-  test: z.boolean(),
+  enableAnnouncementBanner: z.boolean(),
+  enableCmdkTooltip: z.boolean(),
 });
 
 export type EdgeConfig = z.infer<typeof edgeConfigSchema>;
